Added more error handling in crud_operations

+++ models/crud_operations.py
@@ -62,6 +62,9 @@ def delete_product(SKU):
     try:
         product = Product.query.filter_by(SKU=SKU).first()
 
+        if product is None:
+            raise Exception('Product with ' + str(SKU) + ' does not exist')
+
         for batch in product.batches:
             db.session.delete(batch)
 
@@ -95,6 +98,9 @@ def create_product_batch(SKU, producer, batch_quantity, year, month, day):
 
 def read_product_batches(SKU):
     product = Product.query.filter_by(SKU=SKU).first()
+    if product is None:
+        return None
+
     return product.batches
 
 
@@ -120,6 +126,9 @@ def extract_quantity_from_batch(customerID, SKU, requested_quantity):
         requested amount is greated than the stock_quantity
     '''
     product = Product.query.filter_by(SKU=SKU).first()
+
+    if product is None:
+        raise Exception('Product with ' + str(SKU) + ' does not exist')
 
     if product.stock_quantity < requested_quantity:
         raise Exception('Not enough stock_quantity to fill order!')
